Split src and test using tsc --build
Not a huge change in _this_ project, but more cleanly splits things up
and ensures we are actually checking our test files -- by excluding
them previously, we weren't type checking them in the build step.
