+++ src/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "..",
+  "include": ["./**.ts"],
+  "compilerOptions": {
+    "outDir": "../dist"
+  }
+}
