+++ tsconfig.json
@@ -1,14 +1,21 @@
 {
   // see https://www.typescriptlang.org/tsconfig to better understand tsconfigs
-  "include": ["src"],
-  "exclude": ["**/*.test.ts", "**/test/**"],
   "extends": "@tsconfig/node16/tsconfig.json",
+  "references": [
+    {
+      "path": "src"
+    },
+    {
+      "path": "test"
+    }
+  ],
+  "files": [],
   "compilerOptions": {
+    "composite": true,
     "module": "ESNext",
     "declaration": true,
+    "declarationMap": true,
     "sourceMap": true,
-    "rootDir": "./src",
-    "outDir": "./dist",
     // linter checks for common issues
     "noImplicitReturns": true,
     "noFallthroughCasesInSwitch": true,
